refactor(telemetry): split instrumentation API from SDK setup (#538)
common/telemetry bundled the instrumentation helpers AND the full OTel
SDK provider wiring (OTLP trace/metric/log exporters + Prometheus bridge
+ controller-runtime metrics/healthz). Every instrumenting package —
including cni/internal/plugin, linked into the CNI plugin binary that
executes on every container start — dragged all of it in.

Split:
- common/telemetry keeps only the instrumentation API surface:
  attribute keys (attributes.go), EndSpan (span.go), and the gRPC
  stats handlers (grpc.go). No SDK, exporters, Prometheus, or
  controller-runtime deps.
- common/telemetry/setup gets provider construction and shutdown
  plumbing: Setup, SetupTracing, SetupLogs, Config, plus the
  manager-bootstrap helpers ManagerMetricsOptions and CacheSyncChecker.
  Imported only by common/manager (the mains' bootstrap path).

Instrumenting packages keep the common/telemetry import path unchanged;
only common/manager moves to the new setup path.

//cni/cmd/cni dep graph after the split: the Prometheus bridge +
client_golang, otlploggrpc, sdk/log, and controller-runtime are gone.
The remaining otlptracegrpc/otlpmetricgrpc deps come from
cni/internal/telemetry, the plugin's own opt-in (netconf-gated,
force-flushed) exporter path — intentional. Binary: 42.6MB -> 17.9MB.

Fixes #538
