+++ common/telemetry/BUILD.bazel
@@ -5,35 +5,14 @@ go_library(
     srcs = [
         "attributes.go",
         "grpc.go",
-        "healthz.go",
-        "logs.go",
-        "options.go",
         "span.go",
-        "telemetry.go",
-        "tracing.go",
     ],
     importpath = "github.com/bpalermo/aether/common/telemetry",
     visibility = ["//visibility:public"],
     deps = [
-        "@io_k8s_sigs_controller_runtime//:controller-runtime",
-        "@io_k8s_sigs_controller_runtime//pkg/healthz",
-        "@io_k8s_sigs_controller_runtime//pkg/metrics",
-        "@io_k8s_sigs_controller_runtime//pkg/metrics/server",
         "@io_opentelemetry_go_contrib_instrumentation_google_golang_org_grpc_otelgrpc//:otelgrpc",
-        "@io_opentelemetry_go_contrib_instrumentation_runtime//:runtime",
-        "@io_opentelemetry_go_otel//:otel",
         "@io_opentelemetry_go_otel//attribute",
         "@io_opentelemetry_go_otel//codes",
-        "@io_opentelemetry_go_otel//propagation",
-        "@io_opentelemetry_go_otel//semconv/v1.30.0:v1_30_0",
-        "@io_opentelemetry_go_otel_exporters_otlp_otlplog_otlploggrpc//:otlploggrpc",
-        "@io_opentelemetry_go_otel_exporters_otlp_otlpmetric_otlpmetricgrpc//:otlpmetricgrpc",
-        "@io_opentelemetry_go_otel_exporters_otlp_otlptrace_otlptracegrpc//:otlptracegrpc",
-        "@io_opentelemetry_go_otel_exporters_prometheus//:prometheus",
-        "@io_opentelemetry_go_otel_sdk//resource",
-        "@io_opentelemetry_go_otel_sdk//trace",
-        "@io_opentelemetry_go_otel_sdk_log//:log",
-        "@io_opentelemetry_go_otel_sdk_metric//:metric",
         "@io_opentelemetry_go_otel_trace//:trace",
         "@org_golang_google_grpc//stats",
     ],
@@ -41,19 +20,12 @@ go_library(
 
 go_test(
     name = "telemetry_test",
-    srcs = [
-        "logs_test.go",
-        "telemetry_test.go",
-        "tracing_test.go",
-    ],
+    srcs = ["telemetry_test.go"],
     embed = [":telemetry"],
     deps = [
-        "@io_opentelemetry_go_otel//:otel",
         "@io_opentelemetry_go_otel//codes",
         "@io_opentelemetry_go_otel_sdk//trace",
         "@io_opentelemetry_go_otel_sdk//trace/tracetest",
-        "@io_opentelemetry_go_otel_sdk_log//:log",
-        "@io_opentelemetry_go_otel_sdk_metric//:metric",
         "@org_golang_google_grpc//stats",
     ],
 )
