+++ common/telemetry/setup/healthz.go
@@ -1,4 +1,4 @@
-package telemetry
+package setup
 
 import (
 	"errors"
