+++ common/telemetry/setup/logs.go
@@ -1,4 +1,4 @@
-package telemetry
+package setup
 
 import (
 	"context"
