+++ common/telemetry/setup/logs_test.go
@@ -1,4 +1,4 @@
-package telemetry
+package setup
 
 import (
 	"context"
