+++ common/telemetry/setup/tracing.go
@@ -1,4 +1,4 @@
-package telemetry
+package setup
 
 import (
 	"context"
